Time: 4 ms (97.93%), Space: 6 MB (62.86%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -4,7 +4,7 @@ public:
         if(x<0)
             return false;
         int y=x;
-        long long s=0;
+        long s=0;
 
         while(y!=0)
         {
